+++ Internal/Include/ObjectManager.h
@@ -27,7 +27,7 @@ public:
 
 	void Update();
 
-	Object& CreateObject(std::string name, int x, int y, int w, int h, SDL_Texture* texture);
+	Object& CreateObject(std::string name, float x, float y, int w, int h, SDL_Texture* texture);
 
 	void AddObject(Object* object) { mObjects.push_back(object);}
