+++ Internal/Source/ObjectManager.cpp
@@ -25,7 +25,7 @@ void ObjectManager::Update()
 }
 
 
-Object& ObjectManager::CreateObject(std::string name, int x = 0, int y = 0, int w = 0, int h = 0, SDL_Texture* texture = nullptr)
+Object& ObjectManager::CreateObject(std::string name, float x = 0, float y = 0, int w = 0, int h = 0, SDL_Texture* texture = nullptr)
 {
 	Object* object = new Object(name, x, y, w, h, texture);
 	mObjects.push_back(object);
